requirements.txt
pysqlite3-binary
streamlit
pandas
numpy
sentence-transformers
chromadb
google-generativeai